more fixups
