+++ changes/fix_GLT-4245
@@ -1,1 +1,1 @@
-TAT trend report grouping logic
+TAT Trend Report color by gate feature to work correctly with time range grouping
