+++ src/main/resources/templates/tat-trend.html
@@ -14,14 +14,14 @@
       <div id="controlsContainer" class="flex items-center justify-between gap-2 mt-2">
         <div id="groupingButtons" class="flex gap-2">
           <button id="weekButton" data-grouping="week"
-            class="font-inter font-medium text-12 text-black bg-grey-100 px-2 py-1 rounded-md hover:ring-2 ring-green-200 ring-offset-1 flex items-center">Week</button>
+            class="font-inter font-medium text-12 text-black bg-grey-100 px-2 py-1 rounded-md hover:ring-2 ring-green-200 ring-offset-1">Week</button>
           <button id="monthButton" data-grouping="month"
-            class="font-inter font-medium text-12 text-black bg-grey-100 px-2 py-1 rounded-md hover:ring-2 ring-green-200 ring-offset-1 flex items-center">Month</button>
+            class="font-inter font-medium text-12 text-black bg-grey-100 px-2 py-1 rounded-md hover:ring-2 ring-green-200 ring-offset-1">Month</button>
           <button id="quarterButton" data-grouping="fiscalQuarter"
-            class="font-inter font-medium text-12 text-black bg-grey-100 px-2 py-1 rounded-md hover:ring-2 ring-green-200 ring-offset-1 flex items-center">Fiscal
+            class="font-inter font-medium text-12 text-black bg-grey-100 px-2 py-1 rounded-md hover:ring-2 ring-green-200 ring-offset-1">Fiscal
             Quarter</button>
           <button id="yearButton" data-grouping="fiscalYear"
-            class="font-inter font-medium text-12 text-black bg-grey-100 px-2 py-1 rounded-md hover:ring-2 ring-green-200 ring-offset-1 flex items-center">Fiscal
+            class="font-inter font-medium text-12 text-black bg-grey-100 px-2 py-1 rounded-md hover:ring-2 ring-green-200 ring-offset-1">Fiscal
             Year</button>
         </div>
         <div class="flex items-center gap-2">
